requirements.txt
requests>=2.0
python-dotenv>=1.0
Flask>=2.0
gunicorn>=20.1.0
